feat: add configManager setup for environment variable handling
