+++ src/middleware/config-middleware.ts
@@ -1,5 +1,5 @@
 import type { Request, Response, NextFunction } from 'express';
-import { config } from '../config/config-manager';
+import { config } from '../../config-manager/config-manager';
 
 export function configMiddleware(req: Request, res: Response, next: NextFunction) {
   res.locals.config = config;
